Refactor storages to be tightly coupled with vector abstraction (#14)
- Apart from simplifying the code in most places, this allows to keep
the type information of the arbitrary vector on the storage used in a
`ComputeMethod`. Type information is no longer needed for the output
vector using the built-in storages when calling `compute` on available
compute methods.
- Making `ComputeMethod` implementations generic is a little easier
using `ConvertX` traits for both simd and internal vector abstractions,
though they require a `const D: usize` generic.